Add Abilities
This commit defines a list of basic abilities which can be assigned to
roles and which will be used to determine what a user can and cannot do
(based on their role membership).

It is intended that abilities will be named and described in
localization files so that they have meaningful names across all
supported languages. Strictly speaking, it was not necessary to create a
database table for these. A future refactoring may remove this table and
replace it with an enum or something.

```
+-------+      +-------+      +-----------+
| users |*----*| roles |*----*| abilities |
+-------+      +-------+      +-----------+
```

Roles have many abilities and users have many abilities through roles.

Abilities, unlike roles, are not intended to be modified at runtime.
Abilities are enforced via [Pundit](https://github.com/elabs/pundit)
policies in `app/policies`.

NOTE: This commit just makes abilities exist. They don't do anything
yet.