Add Role Administration
Adds the ability to define abilities on roles. Ability names and
descriptions are kept in `config/locales/abilities.en.yml` which means
that they can be I18n'd easily.

Most of the abilities still need names and descriptions in english.

Also adds a sidebar link to the roles index page.