Use Chart.js Gem to Generate Dashboard
The data presented on the dashboard is approximately real. The three
beneficiaries boxes display real and correct data from the database. The
bar charts have real reporting period names on the x-axis. Targets and
actual values are still fake.

Beneficiary, household and services totals are real numbers. The date
is accurate (but in UTC).

Serviced counts for beneficiaries and households are just total counts
of records in the database. Targets and actuals in the bar charts are
random data.